# test: Add perf test to measure the "cost" of chain wakeup