@bhalevy 2024-12-29
@xemul thanks.  This is very important.
BTW, why is the cycles/op stat missing?