@travisdowns 2025-01-03
> BTW, why is the cycles/op stat missing?

There are two related bugs: cycles may always be zero for some benchmarks (https://github.com/scylladb/seastar/issues/2587) and cycles column is missing in .md output (fix: https://github.com/scylladb/seastar/pull/2588).

The first one would be the cause here.