@xemul 2025-01-09
upd:
- fixed (I hope) CI compilation
- normalized results by scale to see the cost of a single continuation