@xemul 2025-01-09
> > it's clear that exception propagation via co-awaits is expensive on _every_ co_await, as it "scales" linearly with the chain depth.
> 
> Makes sense I guess as every co_await in the chain involves a throw-catch and those are the expensive part (high fixed cost, and also a cost per frame unwound though I guess there are very few frames unwound for each throw/catch here).

That's exactly my guess too (Cc @regevran @tchaikov )